+++ Sources/ProjectPlist/ProjectPlist.swift
@@ -3,7 +3,12 @@ import Foundation
 public struct ProjectPlist {
 	typealias ProjectInformation = (bundleId: String, version: String, smallVersion: String)
 
-	public static func createOrUpdateInfoPlist(for projectDirectory: URL, at outputDirectory: URL, for target: String) async throws {
+	public static func createOrUpdateInfoPlist(
+		for projectDirectory: URL,
+		at outputDirectory: URL,
+		for target: String,
+		with pairs: [String: String] = [:]
+	) async throws {
 		let informations = await Self.getInformations(for: projectDirectory, target: target)
 
 		let plist = InfoPlist.at(directory: outputDirectory)
@@ -11,6 +16,9 @@ public struct ProjectPlist {
 		plist.set(name: target)
 		plist.set(version: informations.version)
 		plist.set(smallVersion: informations.smallVersion)
+		pairs.forEach { key, value in
+			plist.setValue(value, forKey: key)
+		}
 
 		try plist.save()
 	}